Refactoring the Javascript Codes
Refactored the codes in the JavaScript file to reduce the lines of codes by introducing the passwordLenght and password and password2 as an empty string variable. The generateNumber() was left intact while the generatePassword() was modified by introducing a for() loop to iterate over the passwordLenght = 15 using i <= passwordLenght. 

Furthermore, within the for() loop two variables where created locally within the for loop to store the loop result and this was equated to the password empty string written at the global level. From there the passwords were equated to the input.value field. 

Hopefully would work on the copy button to allow users click a copy link and have the generated passwords on the clipboard.